worker/requirements.txt
redis
mysql-connector-python
openai-whisper
torch
ffmpeg-python